// src/app/api/familymembers/update/dto.ts
import { z } from 'zod'

export const FamilyMemberUpdateSchema = z.object({
    memberid: z.string().min(1, { message: 'Member ID is required' }),
    familyid: z.number({ required_error: 'Family ID is required' }),
    firstname: z.string().min(1, { message: 'First name is required' }),
    email: z.string().optional().or(z.literal('')).or(z.string().email('Invalid email')),
    colorcode: z.string().min(1, { message: 'Color code is required' }),
    imagedata: z.string().optional(),
    birthdate: z.coerce.date().optional(),
    membersupdatedon: z.string().optional(),
    eventsupdatedon: z.string().optional(),
    externalemail: z.string().optional(),
    emailsystem: z.number().optional(),
    isprivate: z.boolean().optional(),
    localstartdate: z.string().optional(),
    localenddate: z.string().optional(),
    timezone: z.string().optional(),
    offset: z.string().optional(),
    counter: z.number().optional(),
    deleteallappointment: z.boolean().optional(),
    deleteownappointment: z.boolean().optional(),
    createpmtask: z.boolean().optional(),
    pocketmoneyuser: z.boolean().optional(),
    pmtaskapprovedsendconfirmation: z.boolean().optional(),
    hasmasterscheduleaccess: z.boolean().optional(),
    haslocationvisible: z.boolean().optional(),
    hasmemberslocationaccess: z.boolean().optional(),
    locale: z.string().optional(),
})